pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = [
    "Authors Here"
]

[tool.poetry.dependencies]
python = "^3.7"
django = "2.2.13"
gunicorn = "^20.0.4"
djangorestframework = "^3.11.0"
python-decouple = "^3.3"
pillow = "^7.1.2"
django-rest-auth = {extras = ["with_social"], version = "^0.9.5"}
django-cors-headers = "^3.3.0"
celery = "^4.4.3"
redis = "^3.5.3"
django-sendgrid-v5 = "^0.8.1"
django-filter = "^2.2.0"
django-storages = "^1.9.1"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.8.5"
boto3 = "^1.13.24"
django-rest-swagger = "2.1.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.7.9"

[tool.poetry.scripts]
app = ''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
